client_project_manager/client_project_manager/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework.authtoken import views as token_views

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('api.urls')),
    path('api-token-auth/', token_views.obtain_auth_token),
    path('api-auth/', include('rest_framework.urls')),
]